[#25] rename:파일명 제대로 변경
