+++ src/app/profile/page.tsx
@@ -1,4 +1,4 @@
-import Profile from '@/components/mypage/TempProfileInfo';
+import Profile from '@/components/mypage/ProfileInfo';
 import { Metadata } from 'next';
 import React from 'react';
